backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
asyncpg>=0.30.0
python-dotenv>=1.0.0
pydantic>=2.9.0
